+++ app/src/main/AndroidManifest.xml
@@ -31,8 +31,6 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
 
-        <activity android:name=".MainActivity">
-        </activity>
         <activity
             android:name=".RacesActivity"
             android:label="@string/title_activity_races" />
