+++ app/src/main/java/com/trackathon/utn/track_a_thon/HomeActivity.java
@@ -1,6 +1,8 @@
 package com.trackathon.utn.track_a_thon;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
@@ -10,6 +12,7 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -43,6 +46,8 @@ public class HomeActivity extends AppCompatActivity
         ImageView imageView = (ImageView) headerView.findViewById(R.id.navigation_drawer_avatar);
         Picasso.with(headerView.getContext()).load(user.getImageUrl()).into(imageView);
 
+        createElements();
+        createListeners();
     }
 
     @Override
@@ -79,14 +84,14 @@ public class HomeActivity extends AppCompatActivity
 
     @SuppressWarnings("StatementWithEmptyBody")
     @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
+    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
         // Handle navigation view item clicks here.
         int id = item.getItemId();
 
         if (id == R.id.nav_watchers) {
-            // Handle the camera action
+            openActivity(false);
         } else if (id == R.id.nav_trackers) {
-
+            openActivity(true);
         } else if (id == R.id.nav_logout) {
 
         }
@@ -94,5 +99,24 @@ public class HomeActivity extends AppCompatActivity
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
         return true;
+    }
+
+    private Button btnWatchers;
+    private Button btnTrackers;
+
+    private void createElements() {
+        btnWatchers = (Button) findViewById(R.id.btnWatchers);
+        btnTrackers = (Button) findViewById(R.id.btnTrackers);
+    }
+
+    private void createListeners() {
+        btnTrackers.setOnClickListener((View v) -> openActivity(true));
+        btnWatchers.setOnClickListener((View v) -> openActivity(false));
+    }
+
+    private void openActivity (Boolean isTracker) {
+        Intent intent = new Intent(HomeActivity.this, RacesActivity.class);
+        intent.putExtra(TrackatonConstant.IS_TRACKER, isTracker);
+        startActivity(intent);
     }
 }
